requirements.txt
Flask
flask-smorest
sqlalchemy
flask-sqlalchemy
python-dotenv